Parse the output and render it cleanly
Complete the work started in the previous commit. With this, we pull
down the response from the Github API, convert it to plain markdown,
then render the markdown when the Interests page is loaded. This makes
it easier to modify the interests list, at the cost of increased
loading times if the user its the interests url first.